Fix that totaltime records absolute time

+++ src/core/Time.h
@@ -6,12 +6,12 @@ public:
     static void Reset();
     static void Update(double currentTimeSeconds);
 
-    static float GetDeltaTime();
-    static float GetTotalTime();
+    static double GetDeltaTime();
+    static double GetTotalTime();
 
 private:
     static double s_LastTime;
     static double s_TotalTime;
-    static float s_DeltaTime;
+    static double s_DeltaTime;
     static bool s_Initialized;
 };
